Fix ordering of operator compilation for CUDA+MPI.
The CUDA backend needs a list of all boundary tags early on to
build the boundary bitmap. In the CUDA+MPI case, the list of
used tags grows during compilation because MPI rewrites the
operator to include extra boundary terms. This patch makes sure
that the MPI-added tags are available when needed.