Inserted a new method into the RK and AB timesteppers to get an additional constant to calculate the correct timestepp
A visitor pattern like method is used by the discr.dt_factor method to
get a constant factor directly from the timestepper methods to calculate
the timestepp.
As the timestepper needs to be defined before the dt_factor methods
gets called this had to be changed in the example file
example/wave/wave.py. Actually this needs to be changed in every example
file when this feature gets merged with the master!