Space and Time dependent advection implemented
Some Major changes to the pde.py and the exmaple have been made:
Now it is possible to define a space and time dependent advection
function v_advec.
For this purpose the TimeDependentGivenFunction class was used. It had
to be modified making it possible to have more dimensional arrays as
input functions to this class.
This modification is a major modification in data.py
which affects the entire code and will be done in a seperate branch
again.
The changes to the space-time-advection operator and the examples
have sloved down the entire example very strong as another modification
due to calculate the timestepp in the max_eigenvalue method of the
operator. Now several more evaluations of the volume_interpolation have
to be done. It might be worth thinking about a faster implementation.