time and space dependent advection implemented by TimeDependentGivenFunction Interface
Finally the time and space dependent velocity function has been
implemented. It can be switched between a pure space dependent or a time
and space dependent advection (velocity) function.
Additionally the state BC bc_u have been added as time and space
dependet functions with an switchable interfache to the operator. If
they are not assigned a default assignment to u = 0 will be done in the
operator.